+++ index.html
@@ -1,16 +1,14 @@
 <!DOCTYPE html>
 <html lang="ko">
-
-<head>
+  <head>
     <meta charset="UTF-8" />
     <title>자판기</title>
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-</head>
+  </head>
 
-<body>
+  <body>
     <div id="app">
-        <h1>🍿 자판기 🍿</h1>
+      <h1>🍿 자판기 🍿</h1>
     </div>
-</body>
-
+  </body>
 </html>
