+++ src/domains/Coin.ts
